# Relax the constraints of floor_sum