Relax the constraints of floor_sum

Refer to: 
https://github.com/atcoder/ac-library/pull/92

## Benchmark

Current: 
https://atcoder.jp/contests/practice2/submissions/22704612 393ms 
https://atcoder.jp/contests/practice2/submissions/22704628 394ms

New: 
https://atcoder.jp/contests/practice2/submissions/22705926 377ms (with ArgumentError, while true)
https://atcoder.jp/contests/practice2/submissions/22705833 375ms (without ArgumentError, while true)

https://atcoder.jp/contests/practice2/submissions/22705869 457ms (loop method)

## Notes

PR #14 449ms -> 426ms
PR #78 426ms -> 393ms